requirements.txt
langgraph>=0.0.26
langchain-google-genai>=1.0.1
arxiv>=2.1.0
python-dotenv>=1.0.1
tenacity>=8.2.3
pydantic>=2.0.0
